# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for the Azure SRE Demo project.",
        "title": "Azure SRE Demo API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/azure/logs/stream": {
            "get": {
                "description": "Opens a Server-Sent Events (SSE) stream to push live console logs from Azure Log Analytics",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Azure Management"
                ],
                "summary": "Stream Log Analytics Logs",
                "responses": {
                    "200": {
                        "description": "Event stream connected",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/azure/revisions": {
            "get": {
                "description": "Fetches the active revisions and their replicas for the SRE backend container app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Azure Management"
                ],
                "summary": "Get Azure Container Apps Revisions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.RevisionInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.RevisionInfo": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "replicaList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "replicas": {
                    "type": "integer"
                },
                "trafficWeight": {
                    "type": "integer"
                }
            }
        },
...
    }
}